backend/requirements.txt
fastapi
uvicorn
Pillow 
torch
torchvision
openvino 